Rename server folder name
