+++ package.json
@@ -89,18 +89,18 @@
   },
   "scripts": {
     "start-client": "vite",
-    "start-server": "APP_ENV=development nodemon -r dotenv/config --exec ts-node pg-server/index.ts",
+    "start-server": "APP_ENV=development nodemon -r dotenv/config --exec ts-node server/index.ts",
     "build-client": "tsc && vite build",
-    "build-server": "tsc pg-server/index.ts --esModuleInterop --outDir danik-dist",
+    "build-server": "tsc server/index.ts --esModuleInterop --outDir danik-dist",
     "heroku-postbuild": "ts-node ./scripts/build.ts",
     "ts": "tsc --project ./tsconfig.json --noEmit",
     "ts:watch": "tsc --project ./tsconfig.json --watch --noEmit",
-    "lint": "eslint pg-server scripts src --ext .ts,.tsx",
-    "lint-watch": "esw pg-server scripts src --ext .ts,.tsx --watch --color",
+    "lint": "eslint server scripts src --ext .ts,.tsx",
+    "lint-watch": "esw server scripts src --ext .ts,.tsx --watch --color",
     "serve": "node -r dotenv/config danik-dist/index.js",
-    "migrate:make": "knex --knexfile ./pg-server/knexfile.ts migrate:make -x ts",
-    "migrate:latest": "knex --knexfile ./pg-server/knexfile.ts migrate:latest",
-    "migrate:rollback": "knex --knexfile ./pg-server/knexfile.ts migrate:rollback",
+    "migrate:make": "knex --knexfile ./server/knexfile.ts migrate:make -x ts",
+    "migrate:latest": "knex --knexfile ./server/knexfile.ts migrate:latest",
+    "migrate:rollback": "knex --knexfile ./server/knexfile.ts migrate:rollback",
     "create:db:dev": "APP_ENV=development ts-node ./scripts/createDb.ts"
   }
 }
