+++ server/routes/products.ts
