+++ tsconfig.build.json
@@ -15,6 +15,6 @@
     "outDir": "build-server"
   },
   "include": [
-    "pg-server"
+    "server"
   ]
 }
